Add Jupyter notebook with data analysis steps
Included important modules like pandas, matplotlib for data manipulation and visualization. The notebook contains basic steps for data preprocessing and visualization for a given dataset that contains different features of vehicles. The goal of the analysis is to predict the vehicle prices based on the catalog parameters, this will serve as preliminary data exploration before building a predictive model.